pyproject.toml
[tool.poetry]
name = "SentinelScapyScan"
version = "0.1.0"
description = "A modular cybersecurity automation suite for network scanning and service fingerprinting"
authors = ["SentinelScapyScan Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/SentinelScapyScan"
repository = "https://github.com/yourusername/SentinelScapyScan"
keywords = ["security", "network", "scanning", "pentesting", "scapy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: System :: Networking",
]

[tool.poetry.dependencies]
python = "^3.8"
scapy = "^2.5.0"
typer = "^0.9.0"
rich = "^13.7.0"
httpx = "^0.25.0"
jinja2 = "^3.1.2"
cryptography = "^41.0.0"
pyyaml = "^6.0.1"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
pre-commit = "^3.5.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
sentinelscapyscan = "SentinelScapyScan.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=SentinelScapyScan --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["SentinelScapyScan"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
